implementing trace function. fixing sum bugs in execute

+++ Simulator/decode.c
@@ -1,7 +1,8 @@
 //Takes long long, outputs int opcode, int[4] reg addresses, int[2] imm values
 // it masks, shifts and places each value.
 
-int decode(long long input, int regs[], int imm[]) {
+int decode(long long input, int regs[], int imm[]) 
+{
 	imm[1] = (input & 0x7FF) - (input & 0x800);
 	input = (input >> 12);
 	imm[0] = (input & 0x7FF) - (input & 0x800);
@@ -14,5 +15,5 @@ int decode(long long input, int regs[], int imm[]) {
 	input = (input >> 4);
 	regs[0] = (input & 0xF);
 	input = (input >> 4);
-	return input;
+	return (int)input;
 }
